Åpner opp for å migrere saker med historisk ident på barna
